BUGFIX: do not update() SubDomainGrid in ctor
We must not update SubDomainGrid on construction,
as the SubDomainGrid contained in a MultiDomainGrid
cannot update its LevelIndexSets so early, as the
MultiDomainGrid LevelIndexSets are not yet created
when the SubDomainGrid ctor gets called.

update() is now called manually when a SubDomainGrid
is constructed. This is uncritical, as the ctor is
private and we thus control SubDomainGrid creation.