BUGFIX: make sure subIndex entries get marked
markSubIndices was the only functor that only needed
a single data structure to operate. This structure
was thus directly fed into applyToCodims and was
not zipped together with other structures using
fusion::zip(). Unfortunately, this caused the algorithm
to operate on a copy of the index map, thus failing
to really update the index map.

The fix is a littly hacky, as I just zipped the index
map together with the size map, which I simply ignore
later on.