removed the ugly const_cast-hack from applyToCodim
This hack was necessary because the zip_view built
by fusion::zip automatically adds a const to all
involved sequences, making it impossible to modify
the Containers<> object in applyToCodim. Instead of
first adding and then removing the const, the code
now constructs the zip_view by hand, avoiding the
added const altogether.