BUGFIX: make sure the size counters of the new leaf index set are reset
In startSubDomainMarking(), we just copy the old
leaf index set over to the temporary one that will
become the new index set later on. This way, all
subdomain markings are preserved. But: In order to
make sure the index rebuilding works correctly later
on, we have to reset all size containers and clear
the multiIndexMap.