BUGFIX: do not use MultiDomainGrid::hostEntityPointer() in subdomain entity
MultiDomainGrid::hostEntityPointer() was changed to
return a copy instead of a reference in a6fab420d...
While the motivation of this change was very much correct,
it unfortunately broke the way the subdomain entity pointer
accesses the host entity (for geometry access): It
used this method to get at the host entity pointer
embedded in the wrapped multidomain entity pointer,
retrieved its geometry and saved a pointer. Unfortunately,
it now gets a temporary and saves a pointer to its
geometry, which may (and will) vanish very soon.

Fixed by using MultiDomainGrid::hostEntity() instead.