BUGFIX: make sure the geometry wrappers don't acquire stale data in entity copy constructor
During the construction of a subdomain iterator, the
contained multidomain EntityPointer gets constructed
using a copy constructor, which in turn copies the
host EntityPointer (as wanted), but also the geometry
(NOT wanted).

This can lead to invalid memory accesses, as this way
the geometry contains a pointer to the host geometry
of the source multidomain EntityPointer, which might
go out of scope and get destroyed, leaving the copied
geometry pointers dangling.

Fixed by adding explicit copy constructors that only
copy the contained EntityPointer.

Thanks to Nicolas Schwenck for reporting the problem!