BUGFIX: make sure intersections cannot be copied
While it should not be possible to copy intersections,
this rule was unfortunately not enforced by the
compiler because only the copy constructor of the
common base class of Leaf- and LevelIntersections
had restricted visibility, while the copy constructors
of those two were not implemented and thus defaulted
to public visibility.

Fixed by declaring those copy constructors private
and adding Dune::Intersection as a friend class, as
that class has to be able to acces them.