BUGFIX: make sure we don't keep stale level index sets
The current code did not check if a grid adaptation
caused the underlying grid to reduce the number of
levels. In that case, the multidomaingrid update
after the grid adaptation could fail because the code
would try to update its level index sets based on levels
that had disappeared from the host grid.