make protection feature of unsupported codims in dispatchToCodim optional
The dispatchToCodim TMP for dispatching code to a
codim-specific specialization based on a runtime
parameter used to always check if the dispatched-to
codimension was actually supported by the grid. This
feature is now optional, but turned on by default. It
can be turned off by passing an additional template
parameter 'false' to the TMP.

Right now, this is mainly used in the new DataHandle
in the IndexSet class, specifically in the contains()
method.