[EntityPointer] Make sure the SubDomainGrid EntityPointer constructors do not invoke reference casts
The specializations of the SubDomainGrid EntityPointer constructors for the different
types of MultiDomainGrid EntityPointers never matched correctly, because they were
written in terms of the MultiDomainGrid Traits class. As the specializations have to
work for all partition types, the constructors were templated on the PartitionIteratorType,
but the way the specializations were written, this did not work (template parameters can
only be deduced from function arguments and their template parameters, not by just
trying to apply arbitrary parameter values to a struct (pitype) and looking inside those
structs).

This patch fixes the problem by specializing on the actual iterator type as specified in
the grid Traits (a specialization of Dune::EntityIterator). It's ugly, but as long as we have
to cope with the existence of the EntityPointer reference cast operator, it's probably the
only solution.