[Conversions] Make grid entity conversion methods static
The methods used to convert between entities, intersections, iterators
of SubDomainGrid, MultiDomainGrid and the host grid do not really need
an instance of MultiDomainGrid available. This patch makes them static
so that they can also be accessed from places where only the type of the
grid is known, but an instance is not available.