[EntityPointer] Fix MultiDomainGrid EntityPointerWrapper construction for non-standard partitions
The EntityPointerWrapper constructor called from the iterator wrappers is
passed a const reference to the underlying host iterator. Unfortunately, there
are only constructors taking iterators over the default partition (All_Partition).
As there is also a constructor taking a const reference to a host EntityPointer,
constructing a MultiDomainGrid iterator over a non-standard partition would first
invoke the deprecated (and aliasing-rule-violating) cast from the host iterator to
the host EntityPointer and then pass that reference on to the constructor of the
MultiDomainGrid EntityPointer.

With the removal of the deprecated cast from dune-grid in r8024, this mechanism
is now broken - or, depending on your perspective, it always was, and now it's not
possible to cheat around the brokenness anymore... ;-).

Regardless, there are two ways of fixing the issue: We can explicitly specialize
on the interface wrapper for the iterators (as done for a similar issue with SubDomainGrid
in 135523eb), but I would like to avoid doing that for external grids.
So instead, this patch makes the partition type of the iterator available to the
EntityPointer constructor in an extra parameter. That way, the partition type can be used
to correctly specify the type of the host iterator in the second constructor parameter.