Throw an exception on LevelIterator creation on a LeafGridView-only SubDomainGrid
While trying to retrieve a LevelIndexSet from a SubDomainGrid that only supports the
LeafGridView has always resulted in an exception, the creation of a LevelIterator
circumvents this check by directly accessing the underlying data structures.

This patch changes the LevelIterator creation methods of SubDomainGrid to retrieve
the associated LevelIndexSet using the standard interface methods, which will trigger
an exception if LevelGridViews are not supported. This should greatly simplify debugging
when compared to the current situation (SEGFAULT due to null pointer access).