[LevelIterator] Fix stupid copy-and-paste error in HostLevelIterator type deduction
When we try to deduce the host *level* iterator type, we shouldn't extract the
LeafIterator from the host grid...
This issue didn't show up during testing with YaspGrid, as YaspGrid uses the same
iterators for LeafIterators and Leveliterators.

Thanks to Oliver Sander for reporting the issue and providing a reduced example!