[Update] Make sure to purge stale subdomain entries from entities with codim > 0
The subdomain membership of subentities is built from the subdomain membership of
cells in a cumulative fashion during the update process. This will cause problems
if old subdomain memberships are carried over from the previous state and cells
are subsequently removed from a subdomain, as the associated subentities will still
be considered part of that subdomain.

This patch fixes the issue by always clearing out the marking state of all entities
with codim > 0 - it shouldn't be needed anyway, as queries are only possible after
a call to preUpdateSubdomains(), which will update the subdomain memberships to a
consistent state.