[GridView] Add custom GridView implementations that directly construct intersection iterators
WARNING: This patch removes the intersection iterator methods from the entities! The only way to obtain
         intersection iterators now is through a GridView!

The entity interface class has deprecated the i{leaf,level}{begin,end}() methods for obtaining intersection
iterators, they will be removed after the release of Dune 2.3. Unfortunately, this deprecation creates
a rather ugly problem in MultiDomainGrid: As our intersection iterators piggy-back on top of the intersection
iterators of the host grid, the i{leaf,level}{begin,end}() on the entity need to be able to obtain those
iterators. Right now, they call the corresponding methods on the host EntityPointer they store. In order to
solve the deprecation problem in a backwards-compatible way, we would now have to store a reference to the
Grid in our entities and all objects that want to create entities, which turns out to be *really* messy.

So I took the executive decision of breaking backwards compatibility and simply removing the methods on the
entities. I don't think that this will create a lot of breakage because people have been encouraged to use the
GridView methods instead of the Entity methods for a pretty long time.

As the default GridView implementations forward to the Entity methods, that change also required the
introduction of custom GridViews. Those GridViews inherit from the default implementation and only modify
the intersection iterator access methods.