[DynamicSubDomainCountTraits] New storage strategy for runtime-specified subdomain count
The new traits class DynamicSubDomainCountTraits allows the user to specify the
maximum number of subdomains at runtime, as a constructor parameter to the traits.
The traits have to be passed explicitly to the MultiDomainGrid constructor in the
second argument of that constructor.

Apart from being able to specify the number of subdomains at runtime, the new strategy
is identical to the ArrayBasedTraits.