[MultiDomainGrid] Some template magic to exploit a statically known maximum subdomain number
The maximum subdomain number can only be retrieved using a function,
because there are now some traits versions that allow setting this
number at runtime.
On the other hand, it would be nice to still have it available as a
compile-time constant if possible.
This patch does exactly that by making the method static constexpr if
the number is known at compile time, and a regular const member function
otherwise.