[MultiDomainMCMGMapper] Make compatible with new maximum subdomain index querying and dynamic values for the maximum
The patch works by using a std::array if the number is known at compile
time and a std::vector otherwise. In order to correctly handle the
different initialization requirements (the vector needs to allocate
dynamic memory of the right size), the actual storage has been moved to
a mixin base class that contains the variables and handles the
initialization in its constructor and that is specialized for the
different storage schemes.