[MultiDomainMCMGMapper] Use GlobalGeometryTypeIndex and an array to store offsets
The mapper in dune-grid has recently been converted to store the
per-GeometryType offsets in a std::vector indexed by the
GlobalGeometryTypeIndex, which improves lookup times by replacing an
expensive map lookup with a simple array access.

This patch does the same thing for the MultiDomainMCMGMapper and also
gets rid of the separate size variable - that information has been moved
to the end of the offsets array, where it is automatically built as part
of a call to std::partial_sum, converting per-GeometryType sizes to
offsets.

Thanks to Christoph Grüninger for bringing this up!