Revert "Do not pass reference to local objects."
The fix in the original patch introduces a messy bug in Dune 2.3, as
with this patch, the geometry returns a temporary object, but the facade
class  still returns a const reference. That reference now points to the
temporary, which goes out of scope as soon as the facade function
exists. As a result, the user gets a reference to some random stack
memory. Unfortunately, this will normally not trigger a crash; instead,
calculations using the jacobian will simply yield wrong results.

Kudos to Thomas Fetzer for finding the problem.

This reverts commit 62c359d0496c2fb7307026bf8faa1b1fb95c83d5.