[MultiDomainGrid][EntityIterator] Clean up iterators and merge implementations
This patch replaces the separate wrappers for Level- and LeafIterators
with a single class and removes the inheritance from EntityPointer. It
also adds the compatibility conversion and comparison operators with
EntityPointer to tide us over the 2.4 release. Also return entity by
value instead of by refeference to avoid having to store it internally.