[CMake][Gridcheck] Split gridcheck into separate binaries
Having the gridcheck in test-wrapper build a huge mountain of different
combinations of dimension and MultiDomainGrid / SubDomainGrid is not
really helpful to determine where exactly a change has broken
something. At the same time, building that single test takes forever.

So this patch splits test-wrapper into separate binaries for all
combinations of d = 1,2,3 and testing MultiDomainGrid / SubDomainGrid.

With this separation already done, I also went ahead and added separate
tests for checking the grid with / without deprecated APIs.

Finally, the tests are now supposed to compile without generating *any*
warnings, so all gridcheck tests are now compiled with -Werror (in case
of the deprecated API checks, we add -Wno-deprecated-declarations to
avoid triggering an error on the deprecated API calls).